Add files via upload
Music Player App
A simple music player built with JavaScript that allows users to play, pause, shuffle, and skip songs. This app demonstrates basic string and array methods, as well as the use of arrow functions in JavaScript. The project also incorporates the Web Audio API to play songs and manage the playlist.